refactor(data-handler): move performTransition and executeSideEffects onto Project
Removes the containers-importing-commands inversion: calculation-engine
no longer instantiates the Transition command, and the transition command
shrinks to its @write entry point.
